Pull sorting: extra default sort to a module
